refactor: created PokemonList Component
